requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dotenv==0.9.9
fastapi==0.135.1
fonttools==4.62.1
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
kiwisolver==1.5.0
matplotlib==3.10.8
minio==7.2.20
numpy==2.4.3
packaging==26.0
pillow==12.1.1
psycopg2==2.9.11
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
scipy==1.17.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
python-multipart==0.0.20
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.1
